fix: Formatted Pydantic 정의부 분리
